+++ Client/Controller/Controller.vcxproj
@@ -139,9 +139,6 @@
     <ClCompile Include="GameControllerInterface.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Objects\GameObjects\GameObjects.vcxproj">
-      <Project>{f6a0fee2-0f72-44e2-9abc-fa1cf8ba6d4b}</Project>
-    </ProjectReference>
     <ProjectReference Include="..\..\Objects\LogicEngine\LogicEngine.vcxproj">
       <Project>{ace8714c-1dde-4d4d-b947-092acb940888}</Project>
     </ProjectReference>
