+++ Engines/GraphicsEngine/GraphicsEngine.vcxproj
@@ -142,11 +142,14 @@
     <ClCompile Include="MyGraphicsLibraryEngine.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Objects\GameObjects\GameObjects.vcxproj">
-      <Project>{f6a0fee2-0f72-44e2-9abc-fa1cf8ba6d4b}</Project>
+    <ProjectReference Include="..\..\Objects\AbstractObjects\AbstractObjects.vcxproj">
+      <Project>{17253703-7209-43b3-863b-0a8fe5d524eb}</Project>
     </ProjectReference>
     <ProjectReference Include="..\..\Objects\LogicEngine\LogicEngine.vcxproj">
       <Project>{ace8714c-1dde-4d4d-b947-092acb940888}</Project>
+    </ProjectReference>
+    <ProjectReference Include="..\..\Objects\ObjectQueue\ObjectQueue.vcxproj">
+      <Project>{94b05ec3-2fe9-42d9-b765-a322df2c87cd}</Project>
     </ProjectReference>
     <ProjectReference Include="..\RenderTools\RenderTools.vcxproj">
       <Project>{f403ed3b-ea47-415b-b9c9-4ba159b063d8}</Project>
