--- Engines/PhysicsEngine/PhysicalObjectsQueue.cpp
@@ -1,4 +1,0 @@
-#include "PhysicalObjectsQueue.hpp"
-
-PhysicalObjectsQueue::PhysicalObjectsQueue() {}
-PhysicalObjectsQueue::~PhysicalObjectsQueue() {}
