--- Engines/PhysicsEngine/PhysicalObjectsQueue.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include "Objects\GameObjects\AbstractObjectQueue.hpp"
-
-class PhysicalObjectsQueue : public AbstractObjectQueue {
-public:
-	PhysicalObjectsQueue();
-	~PhysicalObjectsQueue();
-};
