+++ Engines/PhysicsEngine/PhysicsEngine.cpp
@@ -1,12 +1,15 @@
 #include "PhysicsEngine.hpp"
+#include "Objects\ObjectQueue\ObjectQueue.hpp"
 
 size_t PhysicsEngine::UpdateInterval = 16667ul;
 
 PhysicsEngine::PhysicsEngine() : m_is_initialized(false) {}
 
-PhysicsEngine::PhysicsEngine(std::function<bool()> const& finishFlagAccess) : PhysicsEngine() {
-	initialize(finishFlagAccess);
+PhysicsEngine::PhysicsEngine(std::function<bool(void)> const& finishFlagAccess, 
+							 ObjectQueue* queue) : PhysicsEngine() {
+	initialize(finishFlagAccess, queue);
 }
+
 PhysicsEngine::~PhysicsEngine() {
 	clean();
 }
@@ -19,30 +22,32 @@ void PhysicsEngine::changeUpdateInterval(size_t microseconds) {
 	UpdateInterval = microseconds;
 }
 
-#include "PhysicalObjectsQueue.hpp"
-void PhysicsEngine::initialize(std::function<bool()> const& finishFlagAccess) {
+void PhysicsEngine::initialize(std::function<bool()> const& finishFlagAccess, ObjectQueue* queue) {
 	m_finish_flag_access = finishFlagAccess;
-	m_queue = new PhysicalObjectsQueue();
+
+	if (queue != nullptr)
+		m_queue = queue;
+	else
+		m_queue = new ObjectQueue();
+
 	m_is_initialized = true;
 }
 void PhysicsEngine::initializeCollisionSystem(GameMap *map) {
 	m_map = map;
 }
-void PhysicsEngine::addObject(AbstractGameObject * object) {
+void PhysicsEngine::addObject(IndependentObject *object) {
 	m_queue->add(object);
 }
-void PhysicsEngine::removeObject(AbstractGameObject * object) {
+void PhysicsEngine::removeObject(IndependentObject *object) {
 	m_queue->remove(object);
 }
 void PhysicsEngine::clean() {
-	if (m_is_initialized && m_queue)
-		delete m_queue;
 	m_is_initialized = false;
 }
 #include <iostream>
 #include <chrono>
 #include <thread>
-#include "Objects\GameObjects\AbstractGameObject.hpp"
+#include "Objects\AbstractObjects\IndependentObject.hpp"
 #include "Shared\GameMode.hpp"
 void PhysicsEngine::loop(bool destroy_engine_after_exit) {
 	GameModeController::physicsLoopIsReady(true);
@@ -50,7 +55,7 @@ void PhysicsEngine::loop(bool destroy_engine_after_exit) {
 		auto next_tick = std::chrono::steady_clock::now() + std::chrono::microseconds(UpdateInterval);
 
 		if (GameModeController::getCurrentGameMode() == GameMode::Normal)
-			m_queue->for_each([this](AbstractGameObject* go) {
+			m_queue->for_each([this](IndependentObject* go) {
 				processForces(go);
 				processMovement(go, m_map);
 			});
