+++ Engines/PhysicsEngine/PhysicsEngine.vcxproj
@@ -115,20 +115,21 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="PhysicalObjectsQueue.cpp" />
     <ClCompile Include="PhysicsEngine.cpp" />
     <ClCompile Include="PhysicsProcessingFunctions.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="PhysicalObjectsQueue.hpp" />
     <ClInclude Include="PhysicsEngine.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Objects\GameObjects\GameObjects.vcxproj">
-      <Project>{f6a0fee2-0f72-44e2-9abc-fa1cf8ba6d4b}</Project>
+    <ProjectReference Include="..\..\Objects\AbstractObjects\AbstractObjects.vcxproj">
+      <Project>{17253703-7209-43b3-863b-0a8fe5d524eb}</Project>
     </ProjectReference>
     <ProjectReference Include="..\..\Objects\LogicEngine\LogicEngine.vcxproj">
       <Project>{ace8714c-1dde-4d4d-b947-092acb940888}</Project>
+    </ProjectReference>
+    <ProjectReference Include="..\..\Objects\ObjectQueue\ObjectQueue.vcxproj">
+      <Project>{94b05ec3-2fe9-42d9-b765-a322df2c87cd}</Project>
     </ProjectReference>
     <ProjectReference Include="..\..\Shared\Shared.vcxproj">
       <Project>{e8a62825-babb-48af-8c6f-c57c08bd594d}</Project>
