+++ Engines/RenderTools/HUD_RenderInfo.hpp
@@ -5,13 +5,13 @@ namespace mgl {
 	class ShaderVariable;
 	class Primitive;
 }
-class AbstractEnergyStorage;
+class EnergyStorage;
 class HUD_RenderInfo : public RenderInfo {
 	mgl::Primitive *m_energy_bar;
-	AbstractEnergyStorage *m_energy_source;
+	EnergyStorage *m_energy_source;
 	float m_current_energy_percent;
 public:
-	HUD_RenderInfo(AbstractEnergyStorage *energy_source);
+	HUD_RenderInfo(EnergyStorage *energy_source);
 	~HUD_RenderInfo();
 
 	virtual void recalculate();
