--- Engines/RenderTools/RenderQueue.cpp
@@ -1,5 +1,0 @@
-#include "RenderQueue.hpp"
-
-RenderQueue::RenderQueue() {}
-
-RenderQueue::~RenderQueue() {}
