--- Engines/RenderTools/RenderQueue.hpp
@@ -1,8 +1,0 @@
-#pragma once
-#include "Objects\GameObjects\AbstractObjectQueue.hpp"
-
-class RenderQueue : public AbstractObjectQueue {
-public:
-	RenderQueue();
-	~RenderQueue();
-};
