+++ Objects/AbstractObjects/ActorObject.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include "Objects\AbstractObjects\IndependentObject.hpp"
+
+class Actor : public IndependentObject {
+public:
+	using IndependentObject::IndependentObject;
+	~Actor() {}
+};
