+++ Objects/AbstractObjects/EquipableItem.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include "Objects\ObjectState\DependentObject.hpp"
+class EquipableItem : public DependentObjectState {
+public:
+	EquipableItem(float mass)
+		: DependentObjectState(mass) {}
+	~EquipableItem() {}
+};
