+++ Objects/EquipableItems/EquipableItems.vcxproj
@@ -18,10 +18,32 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="EnergyStorage.hpp" />
+    <ClInclude Include="FlyEngine.hpp" />
+    <ClInclude Include="Shield.hpp" />
+    <ClInclude Include="Trinket.hpp" />
+    <ClInclude Include="Weapon.hpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="FlyEngine.cpp" />
+    <ClCompile Include="Weapon.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Shared\Shared.vcxproj">
+      <Project>{e8a62825-babb-48af-8c6f-c57c08bd594d}</Project>
+    </ProjectReference>
+    <ProjectReference Include="..\AbstractObjects\AbstractObjects.vcxproj">
+      <Project>{17253703-7209-43b3-863b-0a8fe5d524eb}</Project>
+    </ProjectReference>
+    <ProjectReference Include="..\ObjectState\ObjectState.vcxproj">
+      <Project>{73c2c7ed-90fe-44d1-9402-1056a1624f18}</Project>
+    </ProjectReference>
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
-    <ProjectGuid>{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}</ProjectGuid>
-    <RootNamespace>GameObjects</RootNamespace>
+    <ProjectGuid>{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}</ProjectGuid>
+    <RootNamespace>EquipableItems</RootNamespace>
     <WindowsTargetPlatformVersion>10.0.15063.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -114,24 +136,6 @@
       <OptimizeReferences>true</OptimizeReferences>
     </Link>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClInclude Include="AbstractActors.hpp" />
-    <ClInclude Include="AbstractEquipableItems.hpp" />
-    <ClInclude Include="AbstractGameObject.hpp" />
-    <ClInclude Include="AbstractObjectQueue.hpp" />
-    <ClInclude Include="MainActor.hpp" />
-    <ClInclude Include="ObjectState.hpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="AbstractEquipableItems.cpp" />
-    <ClCompile Include="AbstractObjectQueue.cpp" />
-    <ClCompile Include="MainActor.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Shared\Shared.vcxproj">
-      <Project>{e8a62825-babb-48af-8c6f-c57c08bd594d}</Project>
-    </ProjectReference>
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
