--- Objects/GameObjects/GameObjects.vcxproj.filters
@@ -1,44 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Headers">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;xsd</Extensions>
-    </Filter>
-    <Filter Include="Source">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="AbstractGameObject.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="AbstractObjectQueue.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="AbstractEquipableItems.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="AbstractActors.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="MainActor.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-    <ClInclude Include="ObjectState.hpp">
-      <Filter>Headers</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="AbstractObjectQueue.cpp">
-      <Filter>Source</Filter>
-    </ClCompile>
-    <ClCompile Include="MainActor.cpp">
-      <Filter>Source</Filter>
-    </ClCompile>
-    <ClCompile Include="AbstractEquipableItems.cpp">
-      <Filter>Source</Filter>
-    </ClCompile>
-  </ItemGroup>
-</Project>
