+++ SkyMages.sln
@@ -27,9 +27,17 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RenderTools", "Engines\Rend
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LogicEngine", "Objects\LogicEngine\LogicEngine.vcxproj", "{ACE8714C-1DDE-4D4D-B947-092ACB940888}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GameObjects", "Objects\GameObjects\GameObjects.vcxproj", "{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Shared", "Shared\Shared.vcxproj", "{E8A62825-BABB-48AF-8C6F-C57C08BD594D}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AbstractObjects", "Objects\AbstractObjects\AbstractObjects.vcxproj", "{17253703-7209-43B3-863B-0A8FE5D524EB}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ObjectState", "Objects\ObjectState\ObjectState.vcxproj", "{73C2C7ED-90FE-44D1-9402-1056A1624F18}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Actors", "Objects\Actors\Actors.vcxproj", "{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EquipableItems", "Objects\EquipableItems\EquipableItems.vcxproj", "{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ObjectQueue", "Objects\ObjectQueue\ObjectQueue.vcxproj", "{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -103,14 +111,6 @@ Global
 		{ACE8714C-1DDE-4D4D-B947-092ACB940888}.Release|x64.Build.0 = Release|x64
 		{ACE8714C-1DDE-4D4D-B947-092ACB940888}.Release|x86.ActiveCfg = Release|Win32
 		{ACE8714C-1DDE-4D4D-B947-092ACB940888}.Release|x86.Build.0 = Release|Win32
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Debug|x64.ActiveCfg = Debug|x64
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Debug|x64.Build.0 = Debug|x64
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Debug|x86.ActiveCfg = Debug|Win32
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Debug|x86.Build.0 = Debug|Win32
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Release|x64.ActiveCfg = Release|x64
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Release|x64.Build.0 = Release|x64
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Release|x86.ActiveCfg = Release|Win32
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B}.Release|x86.Build.0 = Release|Win32
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Debug|x64.ActiveCfg = Debug|x64
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Debug|x64.Build.0 = Debug|x64
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Debug|x86.ActiveCfg = Debug|Win32
@@ -119,6 +119,46 @@ Global
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Release|x64.Build.0 = Release|x64
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Release|x86.ActiveCfg = Release|Win32
 		{E8A62825-BABB-48AF-8C6F-C57C08BD594D}.Release|x86.Build.0 = Release|Win32
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Debug|x64.ActiveCfg = Debug|x64
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Debug|x64.Build.0 = Debug|x64
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Debug|x86.ActiveCfg = Debug|Win32
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Debug|x86.Build.0 = Debug|Win32
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Release|x64.ActiveCfg = Release|x64
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Release|x64.Build.0 = Release|x64
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Release|x86.ActiveCfg = Release|Win32
+		{17253703-7209-43B3-863B-0A8FE5D524EB}.Release|x86.Build.0 = Release|Win32
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Debug|x64.ActiveCfg = Debug|x64
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Debug|x64.Build.0 = Debug|x64
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Debug|x86.ActiveCfg = Debug|Win32
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Debug|x86.Build.0 = Debug|Win32
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Release|x64.ActiveCfg = Release|x64
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Release|x64.Build.0 = Release|x64
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Release|x86.ActiveCfg = Release|Win32
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18}.Release|x86.Build.0 = Release|Win32
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Debug|x64.ActiveCfg = Debug|x64
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Debug|x64.Build.0 = Debug|x64
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Debug|x86.ActiveCfg = Debug|Win32
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Debug|x86.Build.0 = Debug|Win32
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Release|x64.ActiveCfg = Release|x64
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Release|x64.Build.0 = Release|x64
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Release|x86.ActiveCfg = Release|Win32
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E}.Release|x86.Build.0 = Release|Win32
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Debug|x64.ActiveCfg = Debug|x64
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Debug|x64.Build.0 = Debug|x64
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Debug|x86.ActiveCfg = Debug|Win32
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Debug|x86.Build.0 = Debug|Win32
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Release|x64.ActiveCfg = Release|x64
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Release|x64.Build.0 = Release|x64
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Release|x86.ActiveCfg = Release|Win32
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC}.Release|x86.Build.0 = Release|Win32
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Debug|x64.ActiveCfg = Debug|x64
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Debug|x64.Build.0 = Debug|x64
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Debug|x86.ActiveCfg = Debug|Win32
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Debug|x86.Build.0 = Debug|Win32
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x64.ActiveCfg = Release|x64
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x64.Build.0 = Release|x64
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x86.ActiveCfg = Release|Win32
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -132,6 +172,10 @@ Global
 		{CB82332B-74C2-4C23-9760-385CADE2CE6A} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
 		{F403ED3B-EA47-415B-B9C9-4BA159B063D8} = {785A6394-7DDC-42FB-A7ED-2898EDC177EE}
 		{ACE8714C-1DDE-4D4D-B947-092ACB940888} = {5865CEC5-B728-447A-824B-4143CCC69C30}
-		{F6A0FEE2-0F72-44E2-9ABC-FA1CF8BA6D4B} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{17253703-7209-43B3-863B-0A8FE5D524EB} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{73C2C7ED-90FE-44D1-9402-1056A1624F18} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{08EBFFE9-3FBC-4778-B56F-EC6212BE743E} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{DFDDC18D-7CDB-4842-94B5-5F7332AC95EC} = {5865CEC5-B728-447A-824B-4143CCC69C30}
+		{94B05EC3-2FE9-42D9-B765-A322DF2C87CD} = {5865CEC5-B728-447A-824B-4143CCC69C30}
 	EndGlobalSection
 EndGlobal
